requirements.txt
fastapi
uvicorn
motor
jinja2
python-multipart
python-jose[cryptography]
passlib[bcrypt]
dnspython
pydantic
email-validator
aiofiles